A new package for text truncation has been added.
This implements a 'clamp' component that allows for the truncation of text by clicking on it. It's especially useful in scenarios where lengthy text needs to be condensed. The component will automatically truncate text depending on the defined height and expand the text upon user interaction. Changes reflected in package-lock.json, index.ts, and added storybook illustration for the same.