Add customizable labels to Truncation component
Providers can now customize the text labels for the 'more' and 'less' buttons within the Truncation component. An overflow property was also added to the ChildrenContainer styled component to hide overflowing content. Additionally, a conditional to only show the clamp label when the content is longer than the defined collapsed height was implemented to avoid unnecessary user interactions. This enhancement improves component flexibility and usability when truncating longer texts.