✨ Add skip initial Accordion animation if shouldDefaultOpen
Enhances accordion component's initial render performance by skipping animations when the component is initially open.

This change introduces a `shouldSkipAnimation` prop to the AccordionHead and AccordionBody components.
It allows disabling the animation on the initial render when the accordion is set to be open by default, reducing initial load times and improving perceived performance.

Additionally, it disables transition on `StyledMotionAccordionBody` to prevent conflicts.