✨ Add UAC group support to Person Finder
Implements the ability to search and display UAC groups within the Person Finder component.

This includes creating a new `PersonFinderSmallItem` component for displaying UAC groups and modifying the `PersonFinderGroup` to render this new component for entries with number IDs (UAC groups).

Updates the `PersonFinderProvider` to handle filtering of the UAC group data. It also removes the dedicated UAC group state and relies solely on the main data store.

The `usePersonFinderItem` hook is also adjusted to provide the appropriate title.

Adds a 'UAC Gruppen' group name to the getGroupName utility function.