✨ Apply cursor repaint fix to input elements
Addresses an issue where the cursor in text inputs on iOS
within the chayns app would sometimes disappear or not render
correctly.

Implements a hook that forces a repaint of the input element
on focus, which resolves the cursor rendering problem.

Removes user specific checks for applying the fix.
Increases amount of repaints triggered.